requirements.txt
bpython
beautifulsoup4
urllib3
mechanize
pillow
pyimgur
hachoir-core
hachoir-metadata
hachoir-parser